feat: TextInput 상태 메시지 로직 수정 - rename: components type 폴더 생성
