+++ src/hooks/useFocus.ts
@@ -1,12 +1,11 @@
-import {ITextInputTypes} from '@/types/textInput';
+import {ITextInput} from '@/types/ui';
 import {useState} from 'react';
 
 /**
  * focus에 따른 색상 여부 전달
  */
 const useFocus = () => {
-  const [type, setType] = useState<ITextInputTypes>('default');
-
+  const [type, setType] = useState<ITextInput>('default');
   /**
    * Focus 처리합니다.
    */
@@ -17,7 +16,6 @@ const useFocus = () => {
   /**
    * maxLength input 에러 처리합니다.
    */
-
   const onError = () => {
     setType('caution');
   };
