--- src/types/button.ts
@@ -1,1 +1,0 @@
-export type IVariant = 'primary' | 'secondary' | 'default';
