--- src/types/checkBox.ts
@@ -1,1 +1,0 @@
-export type ICheckBox = 'on' | 'off' | 'indeterminate';
