--- src/types/chip.ts
@@ -1,1 +1,0 @@
-export type IChip = 'primary' | 'default';
