--- src/types/textInput.ts
@@ -1,2 +1,0 @@
-export type ITextInput = 'default' | 'writed' | 'caution';
-export type ITextInputTypes = 'default' | 'writed' | 'caution' | 'active';
