+++ src/types/ui/index.ts
@@ -1,0 +1,11 @@
+//버튼 타입 지정
+export type IVariant = 'primary' | 'secondary' | 'default';
+
+// 체크박스 타입 지정
+export type ICheckBox = 'on' | 'off' | 'indeterminate';
+
+// chip 타입 지정
+export type IChip = 'primary' | 'default';
+
+// textInput 타입 지정
+export type ITextInput = 'default' | 'writed' | 'caution' | 'active';
